requirements.txt
pydantic
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
argon2-cffi
firebase-admin
requests
python-multipart
jinja2
python-jose
itsdangerous
starlette
argon2